# docs/swagger.yaml
basePath: /
definitions:
...
  task.taskForm:
    properties:
      contributor:
        type: string
      description:
        type: string
      end_date:
        type: integer
      name:
        type: string
      project_id:
        type: string
      start_date:
        type: integer
    type: object
info:
  contact: {}
  description: Proman Backend API
  title: Proman Backend
paths:
...
  /api/me/projects:
    get:
      consumes:
      - application/json
      operationId: my-projects
      parameters:
      - description: Search by nama or description
        in: query
        name: q
        type: string
      - description: Search by status
        enum:
        - active
        - completed
        - pending
        - cancelled
        in: query
        name: status
        type: string
      - description: Start date
        in: query
        name: start
        type: string
      - description: End date
        in: query
        name: end
        type: string
      - description: Sort
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - description: Page number pagination
        in: query
        name: page
        type: integer
      - description: Limit pagination
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get my projects
      tags:
      - Me Project
...
  /api/me/tasks:
    get:
      consumes:
      - application/json
      operationId: my-tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get my tasks
      tags:
      - Me Task
  /api/project:
    post:
      consumes:
      - application/json
      operationId: create-project
      parameters:
      - description: Project name
        in: formData
        name: name
        required: true
        type: string
      - description: Project description
        in: formData
        name: description
        required: true
        type: string
      - description: Project start date
        in: formData
        name: start_date
        required: true
        type: integer
      - description: Project end date
        in: formData
        name: end_date
        required: true
        type: integer
      - description: Project contributor
        in: formData
        name: contributor
        required: true
        type: string
      - description: Project type
        enum:
        - frontend
        - backend
        - mobile
        - desktop
        - monitor
        - tool
        - etc
        in: formData
        name: type
        required: true
        type: string
      - description: Project logo
        in: formData
        name: logo
        type: file
      - description: Project attachments
        in: formData
        name: attachments
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create project
      tags:
      - Project
  /api/project/{id}:
    delete:
      consumes:
      - application/json
      operationId: delete-project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete project by id
      tags:
      - Project
    get:
      consumes:
      - application/json
      operationId: get-project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get project by id
      tags:
      - Project
  /api/project/count:
    get:
      consumes:
      - application/json
      operationId: count-project
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get project count
      tags:
      - Project
  /api/project/count/type:
    get:
      consumes:
      - application/json
      operationId: count-project-type
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get project count by type
      tags:
      - Project
  /api/projects:
    get:
      consumes:
      - application/json
      operationId: list-project
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get list of project
      tags:
      - Project
...
  /api/task:
    post:
      consumes:
      - application/json
      operationId: task-create
      parameters:
      - description: Task data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/task.taskForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create task
      tags:
      - Task
  /api/task/{id}:
    delete:
      consumes:
      - application/json
      operationId: task-delete
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete task
      tags:
      - Task
    get:
      consumes:
      - application/json
      operationId: task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get task by ID
      tags:
      - Task
  /api/task/count:
    get:
      consumes:
      - application/json
      operationId: task-count
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get task count
      tags:
      - Task
  /api/task/overview:
    get:
      consumes:
      - application/json
      operationId: task-overview
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get task overview
      tags:
      - Task
  /api/task/status:
    get:
      consumes:
      - application/json
      operationId: task-list-status
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get task list by status
      tags:
      - Task
  /api/tasks:
    get:
      consumes:
      - application/json
      operationId: tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get tasks
      tags:
      - Task
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"